Missing header

+++ src/libUtil/include/Histo3d.h
@@ -74,7 +74,7 @@ class Histo3d : public HistogramBase {
         void plot(std::string filename, std::string dir = "");
 
     private:
-        double *data;
+        uint16_t *data;
 
         double underflow;
         double overflow;
